backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
chromadb==0.4.22
openai==1.12.0
google-generativeai==0.8.3
pymupdf==1.23.22
httpx==0.26.0
python-dotenv==1.0.0
aiofiles==23.2.1
numpy<2.0
